% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergent.transitions_function.R
\name{check.divergent.transitions}
\alias{check.divergent.transitions}
\title{Counts the number of divergent transitions in a fitted stan model.}
\usage{
check.divergent.transitions(fit, model_name = NULL)
}
\arguments{
\item{fit}{Stan model fit object.}

\item{model_name}{Optional model name identifier.}
}
\value{
Specified model name with the number of divergent transitions
}
\description{
Counts the number of divergent transitions in a fitted stan model.
}
