% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulating.test.data.exp_functions.R
\name{exponential.model.run}
\alias{exponential.model.run}
\title{Generates disease prevalence assuming exponential growth}
\usage{
exponential.model.run(days = 365, I0 = 1e-04, beta0 = 0.06, gamma = 0.04)
}
\arguments{
\item{days}{Number of days to simulate disease growth.}

\item{I0}{Initial disease prevalence. Default = 0.0001.}

\item{beta0}{Base transmission rate. Default = 0.06.}

\item{gamma}{Recovery rate. Default = 0.04.}
}
\value{
A dataframe of model results at each timepoint. Includes: Time = day; I_t = prevalence; R0_t = reproduction number .
}
\description{
Generates disease prevalence assuming exponential growth to model true prevalence at the start of an epidemic.
}
\examples{
if (interactive()) {

exp_mod_result <- exponential.model.run(days=365, I0 = 0.0001,  beta0 = 0.06, gamma = 0.04)

# Plot the simulated cases over time
exp_mod_result \%>\%
ggplot() +
 geom_line(aes(x=time, y=I_t)) +
 theme_classic() +
 labs(x="Days", y="Prevalence")

# Plot the reproduction number over time
exp_mod_result \%>\%
 ggplot() +
  geom_line(aes(x=time, y=R0_t)) +
  theme_classic() +
  labs(x="Days", y="Reproduction number (Rt)") +
  scale_y_continuous(labels = function(x) sprintf("\%.2f", x))

}
}
