% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.sim.LC.prev.time_functionR.R
\name{prev.time.plots}
\alias{prev.time.plots}
\title{Plots inferred weekly prevalence from run.LC.model with covariates = c("Time")}
\arguments{
\item{run.LC.model_output}{Returned run.LC.model() model object.}

\item{sim_time_data}{Returned output from sim.test.data.time() if simulated data were used. Allows plotting of true prevalence over time alongside inferred. Default = NULL.}

\item{time_resolution}{Time resolution of test data used in run.LC.model(). Either "days" or "weeks". Default = "days".}
}
\value{
Stan model fit and various summary outputs:
\describe{
\item{weekly_inferred_prev_plot}{Plot of weekly inferred prevalence from the LC model which adjusts test data for test error.}
\item{weekly_prev_plot}{Plot of weekly inferred prevalence from the LC model which adjusts test data for test error, with plotted test positivity (apparent prevalence), and true prevalence if simulated}
\item{weekly_data}{Data aggregated to weekly}
}
}
\description{
Plots inferred weekly prevalence from run.LC.model with covariates = c("Time"). If sim.test.data.time() was used to simulate true prevalence and test results,
this function output can be provided to the sim_time_data argument to compare inferred prev with true prevalence.
}
