% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulating.test.data.time_functions.R
\name{run.sir.model}
\alias{run.sir.model}
\title{Generates and solves an SIR model with seasonal forcing}
\usage{
run.sir.model(
  years = 50,
  N = 1,
  init = list(init_S = 0.99, init_I = 0.01, init_R = 0),
  params = list(beta0 = NULL, desired_R0 = 2.5, beta1 = 0.07, phi = 1.5, gamma = 0.03,
    omega = 0.001)
)
}
\arguments{
\item{years}{Number of years to run SIR. Default = 50.}

\item{N}{Population size. Default = 1.}

\item{init}{List of initial state values. Default = init_S = 0.99; init_I = 0.01; init_R = 0.}

\item{params}{list of SIR paramters. Default = beta0 = NULL, desired_R0 = 2.5, beta1 = 0.07, phi = 1.5, gamma = 0.03, omega = 0.001. If beta0 is null it is calculated based on desired_R0.}
}
\value{
SIR model results including beta_t and R0_t and Re_t
}
\description{
Generates and solves an SIR model with seasonal forcing to model true prevalence over the course of a seasonal disease with a single peak.
}
