% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.sim.vs.LC_function.R
\name{sim.result.plot}
\alias{sim.result.plot}
\title{Plots LC model simulation results.}
\arguments{
\item{sim_stan_results}{Named output from run.sims.LC().}
}
\value{
Plots of inferred parameters against simulated parameters. Returns 3 plots:
#' \describe{
\item{1) Prevalence}{LC model inferred prevalence (mean across chains, mean across iterations)) vs simulated prevalence;}
\item{2) Sensitivity}{LC model inferred test sensitivities (mean across chains, median across iterations) vs simulated test sensitivities;}
\item{3) Specificity}{LC model inferred test specificities (mean across chains, median across iterations) vs simulated test specificities;}
}
}
\description{
Plots LC model simulation results generated from run.sims.LC() to evaluate model performance at different parameter values.
Compares LC model inferred prevalence, test sensitivities and test specificities against simulated parameters.
Will plot across any simulated any value of the probability of the test occurring (sim_prob),
so if multiple p_performed were simulated in run.sims.LC(), these will need to be pre-filtered to plot each sim_prob paramter value one at a time.
}
\examples{
if (interactive()) {
sim_results <- run.sims.LC(num_tests = 4,
                       prev_vec = c(0.1, 0.2),
                       spec_vec = c(1, 0.98),
                       sens_vec = c(1),
                       p_performed_vec= c(1),
                       sim_size = 100,
                       chains = 2,
                       data_ID = "sims")

sim.result.plots(sim_results)
}

}
