% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Divergent.transitions_function.R
\name{check.divergent.transitions}
\alias{check.divergent.transitions}
\title{Counts the number of divergent transitions in a fitted stan model.}
\usage{
check.divergent.transitions(fit, model_name = NULL)
}
\arguments{
\item{fit}{Stan model fit object.}

\item{model_name}{Optional model name identifier.}
}
\value{
Specified model name with the number of divergent transitions
}
\description{
Counts the number of divergent transitions in a fitted stan model.
}
\examples{
if (interactive()) {
# Simulate data
sim_data <- sim.test.data(sim_size = 100)

# Run LC model using simulated test results
fit <- run.LC.model(sim_data$test_results, num_tests = 4,
                      data_ID = "sim", model_name = "basic_sim",
                      dependency_groups = NULL, covariates = NULL,
                      iter=1000, chains=4, warmup=500)

check.divergent.transitions(fit$stan_fit, model_name = "Test_model")
}

}
