% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate.stan.LCmodel_function.R
\name{generate.stan.model}
\alias{generate.stan.model}
\title{Creates stan model code for different versions of the LC model that can be run in R using rstan.}
\usage{
generate.stan.model(
  num_tests,
  include_time = FALSE,
  include_delay = FALSE,
  dependency_groups = list()
)
}
\arguments{
\item{num_tests}{Number of tests in data.}

\item{include_time}{Logical indicating whether time should be included in the model. Default = FALSE.}

\item{include_delay}{Logical indicating whether delay until testing should be included in the model. Default = FALSE.}

\item{dependency_groups}{A list of vectors specifying dependencies between tests. All test numbers that are not independent of each other are specified in a single vector. Default = NULL.}
}
\value{
Stan LC model code.
}
\description{
Creates stan model code for different versions of the LC model that can be run in R using rstan.
}
