% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Running.LC.model.dynamic_function.R
\name{run.LC.model}
\alias{run.LC.model}
\title{Uses test result data to run a bayesian latent class stan model to infer true disease prevalence whilst accounting for test error}
\usage{
run.LC.model(
  data,
  num_tests,
  test_names_defined = NULL,
  data_ID = NULL,
  dependency_groups = list(),
  iter = 1000,
  chains = 4,
  warmup = 500,
  stan_arg = list(),
  covariates = NULL,
  prior_list = NULL,
  n_samples = iter,
  model_name = NULL
)
}
\arguments{
\item{data}{A dataframe of binary test results with a column for each test and row for each individual. If delay until test covariates are to be included in the model, a column with delay data for each test should be included after all of the test result data columns, in the same test order as the test result data.}

\item{num_tests}{A numeric value for the number of tests.}

\item{test_names_defined}{A character vector of identifiers / names for each test, given in the same order as the data. Default = NULL.}

\item{data_ID}{An optional name identifier for the dataset used. Default = NULL.}

\item{dependency_groups}{A list of vectors specifying dependencies between tests. All test numbers that are not independent of each other are specified in a single vector. Default = NULL.}

\item{iter}{The number of iterations for the stan model. Default = 1000.}

\item{chains}{The number of chains for the stan model. Default = 4.}

\item{warmup}{The number of warmup iterations for the stan model. Default = 500.}

\item{stan_arg}{Optional extra arguments to pass to the rstan::sampling function. Default = NULL.}

\item{covariates}{Optional covariate to include in the model. Either "Time" or "Delay". If specified, relevant individual-level data must be provided in data. Default = NULL.}

\item{prior_list}{Optional list of prior distributions-specified to generate in R- to plot with posteriors.}

\item{n_samples}{If prior_list provided, number of times to samples from the specified prior distributions. Default = 1000.}

\item{model_name}{Optional name of model. Default = NULL.}
}
\value{
Stan model fit and various summary outputs.
}
\description{
Uses test result data to run a bayesian latent class stan model to infer true disease prevalence whilst accounting for test error
}
