% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulating.test.data.time_functions.R
\docType{data}
\name{run.sir.model}
\alias{run.sir.model}
\alias{params}
\title{Generates and solves an SIR model with seasonal forcing}
\format{
An object of class \code{list} of length 6.
}
\usage{
params
}
\arguments{
\item{years}{Number of years to run SIR. Default = 50.}

\item{N}{Population size. Default = 1.}

\item{init}{List of initial state values. Default = init_S = 0.99; init_I = 0.01; init_R = 0.}

\item{params}{list of SIR paramters. Default = beta0 = NULL, desired_R0 = 2.5, beta1 = 0.07, phi = 1.5, gamma = 0.03, omega = 0.001. If beta0 is null it is calculated based on desired_R0.}
}
\value{
SIR model results including beta_t and R0_t and Re_t
}
\description{
Generates and solves an SIR model with seasonal forcing
}
\keyword{datasets}
